requirements.txt
telethon==1.40.0
tqdm==4.66.1
aiofiles==23.2.1
python-dotenv==1.0.0